Refactor remuneration files and generate remuneration preview route and controller
