+++ src/components/employee/remuneration/remuneration-details.js
@@ -1,0 +1,11 @@
+/**
+ * Generate an array of remunerations details following the rules defined by the labor union.
+ * @param {Object} employee Employee object
+ */
+module.exports.calculate = (employee) => {
+    const details = [];
+
+    
+
+    return details;
+}
